fix(spec): preserve software callee matching
